+++ Realm/RLMRealm.h
@@ -28,6 +28,9 @@
  */
 typedef void(^RLMAsyncOpenRealmCallback)(RLMRealm * _Nullable realm, NSError * _Nullable error);
 
+/// The Id of the asynchronous transaction.
+typedef unsigned RLMAsyncTransactionId;
+
 NS_ASSUME_NONNULL_BEGIN
 
 /**
@@ -516,6 +519,151 @@ typedef void (^RLMNotificationBlock)(RLMNotification notification, RLMRealm *rea
  @return Whether the transaction succeeded.
  */
 - (BOOL)transactionWithoutNotifying:(NSArray<RLMNotificationToken *> *)tokens block:(__attribute__((noescape)) void(^)(void))block error:(NSError **)error;
+
+/**
+ Indicates if the Realm is currently performing async write operations.
+ This becomes YES following a call to `beginAsyncWriteTransaction`,
+ `commitAsyncWriteTransaction`, or `asyncTransactionWithBlock:`, and remains so
+ until all scheduled async write work has completed.
+ 
+ @warning If this is `YES`, closing or invalidating the Realm will block until scheduled work has completed.
+ */
+@property (nonatomic, readonly) BOOL isPerformingAsynchronousWriteOperations;
+
+/**
+ Begins an asynchronous write transaction.
+ This function asynchronously begins a write transaction on a background
+ thread, and then invokes the block on the original thread or queue once the
+ transaction has begun. Unlike `beginWriteTransaction`, this does not block the
+ calling thread if another thread is current inside a write transaction, and
+ will always return immediately.
+ Multiple calls to this function (or the other functions which perform
+ asynchronous write transactions) will queue the blocks to be called in the
+ same order as they were queued. This includes calls from inside a write
+ transaction block, which unlike with synchronous transactions are allowed.
+ 
+ @param block The block containing actions to perform inside the write transaction.
+        `block` should end by calling `commitAsyncWriteTransaction`,
+        `commitWriteTransaction` or `cancelWriteTransaction`.
+        Returning without one of these calls is equivalent to calling `cancelWriteTransaction`.
+ 
+ @return An id identifying the asynchronous transaction which can be passed to
+         `cancelAsyncTransaction:` prior to the block being called to cancel
+         the pending invocation of the block.
+ */
+- (RLMAsyncTransactionId)beginAsyncWriteTransaction:(void(^)(void))block;
+
+/**
+ Asynchronously commits a write transaction.
+ The call returns immediately allowing the caller to proceed while the I/O is
+ performed on a dedicated background thread. This can be used regardless of if
+ the write transaction was begun with `beginWriteTransaction` or
+ `beginAsyncWriteTransaction`.
+ 
+ @param completionBlock A block which will be called on the source thread or
+                        queue once the commit has either completed or failed
+                        with an error.
+ 
+ @param isGroupingAllowed If `YES`, multiple sequential calls to
+                         `commitAsyncWriteTransaction:` may be batched together
+                         and persisted to stable storage in one group. This
+                         improves write performance, particularly when the
+                         individual transactions being batched are small. In the
+                         event of a crash or power failure, either all of the
+                         grouped transactions will be lost or none will, rather
+                         than the usual guarantee that data has been persisted as
+                         soon as a call to commit has returned.
+ 
+ @return An id identifying the asynchronous transaction commit can be passed to
+         `cancelAsyncTransaction:` prior to the completion block being called
+         to cancel the pending invocation of the block. Note that this does
+         *not* cancel the commit itself.
+*/
+- (RLMAsyncTransactionId)commitAsyncWriteTransaction:(nullable void(^)(NSError *))completionBlock
+                                   isGroupingAllowed:(BOOL)isGroupingAllowed;
+
+/**
+ Asynchronously commits a write transaction.
+ The call returns immediately allowing the caller to proceed while the I/O is
+ performed on a dedicated background thread. This can be used regardless of if
+ the write transaction was begun with `beginWriteTransaction` or
+ `beginAsyncWriteTransaction`.
+ 
+ @param completionBlock A block which will be called on the source thread or
+                        queue once the commit has either completed or failed
+                        with an error.
+ 
+ @return An id identifying the asynchronous transaction commit can be passed to
+         `cancelAsyncTransaction:` prior to the completion block being called
+         to cancel the pending invocation of the block. Note that this does
+         *not* cancel the commit itself.
+*/
+- (RLMAsyncTransactionId)commitAsyncWriteTransaction:(void(^)(NSError *))completionBlock;
+
+/**
+ Asynchronously commits a write transaction.
+ The call returns immediately allowing the caller to proceed while the I/O is
+ performed on a dedicated background thread. This can be used regardless of if
+ the write transaction was begun with `beginWriteTransaction` or
+ `beginAsyncWriteTransaction`.
+ 
+ @return An id identifying the asynchronous transaction commit can be passed to
+         `cancelAsyncTransaction:` prior to the completion block being called
+         to cancel the pending invocation of the block. Note that this does
+         *not* cancel the commit itself.
+*/
+- (RLMAsyncTransactionId)commitAsyncWriteTransaction;
+
+/**
+ Cancels a queued block for an asynchronous transaction.
+ This can cancel a block passed to either an asynchronous begin or an
+ asynchronous commit. Canceling a begin cancels that transaction entirely,
+ while canceling a commit merely cancels the invocation of the completion
+ callback, and the commit will still happen.
+ Transactions can only be canceled before the block is invoked, and calling
+ `cancelAsyncTransaction:` from within the block is a no-op.
+ 
+ @param asyncTransactionId A transaction id from either `beginAsyncWriteTransaction:` or `commitAsyncWriteTransaction:`.
+*/
+- (void)cancelAsyncTransaction:(RLMAsyncTransactionId)asyncTransactionId;
+
+/**
+ Asynchronously performs actions contained within the given block inside a
+ write transaction.
+ The write transaction is begun asynchronously as if calling
+ `beginAsyncWriteTransaction:`, and by default the transaction is commited
+ asynchronously after the block completes. You can also explicitly call
+ `commitWriteTransaction` or `cancelWriteTransaction` from within the block to
+ synchronously commit or cancel the write transaction.
+ 
+ @param block The block containing actions to perform.
+ 
+ @param completionBlock A block which will be called on the source thread or
+                        queue once the commit has either completed or failed
+                        with an error.
+
+ @return An id identifying the asynchronous transaction which can be passed to
+         `cancelAsyncTransaction:` prior to the block being called to cancel
+         the pending invocation of the block.
+*/
+- (RLMAsyncTransactionId)asyncTransactionWithBlock:(void(^)(void))block onComplete:(nullable void(^)(NSError *))completionBlock;
+
+/**
+ Asynchronously performs actions contained within the given block inside a
+ write transaction.
+ The write transaction is begun asynchronously as if calling
+ `beginAsyncWriteTransaction:`, and by default the transaction is commited
+ asynchronously after the block completes. You can also explicitly call
+ `commitWriteTransaction` or `cancelWriteTransaction` from within the block to
+ synchronously commit or cancel the write transaction.
+ 
+ @param block The block containing actions to perform.
+ 
+ @return An id identifying the asynchronous transaction which can be passed to
+         `cancelAsyncTransaction:` prior to the block being called to cancel
+         the pending invocation of the block.
+*/
+- (RLMAsyncTransactionId)asyncTransactionWithBlock:(void(^)(void))block;
 
 /**
  Updates the Realm and outstanding objects managed by the Realm to point to the